requirements.txt
streamlit
qdrant-client
openai
fastembed
pypdf
python-dotenv
tqdm
tenacity
python-docx


